🎨 ComponentLibrary 완전 모듈화 완료
✨ 새로운 구조:
ComponentLibrary/
├── thumbnails/          # 컴포넌트별 썸네일
│   ├── ButtonThumbnail.jsx
│   ├── TextThumbnail.jsx
│   ├── AttendThumbnail.jsx
│   ├── DdayThumbnail.jsx
│   ├── CalendarThumbnail.jsx
│   ├── BankAccountThumbnail.jsx
│   ├── CommentThumbnail.jsx
│   ├── ImageThumbnail.jsx
│   ├── MapThumbnail.jsx
│   ├── WeddingContactThumbnail.jsx
│   ├── DefaultThumbnail.jsx
│   └── index.js
├── components/          # UI 컴포넌트들
│   ├── ComponentCard.jsx
│   ├── LibraryHeader.jsx
│   ├── SearchBar.jsx
│   ├── ComponentGrid.jsx
│   ├── ToggleButton.jsx
│   └── DeploySection.jsx
├── hooks/               # 커스텀 훅
│   └── useDeploy.js
└── utils/               # 유틸리티 (향후 확장)

🔧 개선사항:
- 완전한 관심사 분리
- 재사용 가능한 컴포넌트 구조
- 커스텀 훅으로 로직 분리
- 썸네일 개별 수정 가능
- 확장성 및 유지보수성 대폭 향상
