+++ my-web-builder/apps/frontend/src/pages/NoCodeEditor/ComponentLibrary/thumbnails/ButtonThumbnail.jsx
@@ -1,0 +1,22 @@
+import React from 'react';
+
+function ButtonThumbnail() {
+  return (
+    <div style={{
+      width: 60, 
+      height: 24, 
+      background: '#3B4EFF', 
+      color: '#fff',
+      borderRadius: 4, 
+      display: 'flex', 
+      alignItems: 'center', 
+      justifyContent: 'center',
+      fontSize: 10, 
+      fontWeight: 'bold'
+    }}>
+      Button
+    </div>
+  );
+}
+
+export default ButtonThumbnail;
