+++ my-web-builder/apps/frontend/src/pages/NoCodeEditor/ComponentLibrary/thumbnails/LinkThumbnail.jsx
@@ -1,0 +1,16 @@
+import React from 'react';
+
+function LinkThumbnail() {
+  return (
+    <div style={{
+      fontSize: 12, 
+      color: '#3B4EFF', 
+      textDecoration: 'underline',
+      textAlign: 'center'
+    }}>
+      Link Text
+    </div>
+  );
+}
+
+export default LinkThumbnail;
