+++ my-web-builder/apps/frontend/src/pages/NoCodeEditor/ComponentLibrary/thumbnails/MapThumbnail.jsx
@@ -1,0 +1,22 @@
+import React from 'react';
+
+function MapThumbnail() {
+  return (
+    <div style={{
+      width: 70, 
+      height: 50, 
+      background: '#e8f5e8', 
+      border: '1px solid #c8e6c9',
+      borderRadius: 4, 
+      display: 'flex', 
+      alignItems: 'center', 
+      justifyContent: 'center',
+      fontSize: 16,
+      color: '#2e7d32'
+    }}>
+      🗺️
+    </div>
+  );
+}
+
+export default MapThumbnail;
