+++ my-web-builder/apps/frontend/src/pages/NoCodeEditor/ComponentLibrary/thumbnails/TextThumbnail.jsx
@@ -1,0 +1,17 @@
+import React from 'react';
+
+function TextThumbnail() {
+  return (
+    <div style={{
+      fontSize: 12, 
+      color: '#333', 
+      fontWeight: 'normal',
+      textAlign: 'center', 
+      lineHeight: 1.2
+    }}>
+      Sample Text
+    </div>
+  );
+}
+
+export default TextThumbnail;
