+++ my-web-builder/apps/frontend/src/pages/NoCodeEditor/ComponentLibrary/thumbnails/index.js
@@ -1,0 +1,54 @@
+import ButtonThumbnail from './ButtonThumbnail';
+import TextThumbnail from './TextThumbnail';
+import LinkThumbnail from './LinkThumbnail';
+import AttendThumbnail from './AttendThumbnail';
+import DdayThumbnail from './DdayThumbnail';
+import CalendarThumbnail from './CalendarThumbnail';
+import BankAccountThumbnail from './BankAccountThumbnail';
+import CommentThumbnail from './CommentThumbnail';
+import ImageThumbnail from './ImageThumbnail';
+import MapThumbnail from './MapThumbnail';
+import WeddingContactThumbnail from './WeddingContactThumbnail';
+import DefaultThumbnail from './DefaultThumbnail';
+
+export const ThumbnailComponents = {
+  button: ButtonThumbnail,
+  text: TextThumbnail,
+  link: LinkThumbnail,
+  attend: AttendThumbnail,
+  dday: DdayThumbnail,
+  'd-day': DdayThumbnail, // kebab-case alias
+  calendar: CalendarThumbnail,
+  bankAccount: BankAccountThumbnail,
+  'bank-account': BankAccountThumbnail, // kebab-case alias
+  comment: CommentThumbnail,
+  image: ImageThumbnail,
+  map: MapThumbnail,
+  mapInfo: MapThumbnail, // map_info uses same thumbnail
+  'map_info': MapThumbnail, // underscore alias
+  weddingContact: WeddingContactThumbnail,
+  'wedding-contact': WeddingContactThumbnail, // kebab-case alias
+  weddingInvite: WeddingContactThumbnail, // reuse wedding contact
+  'wedding-invite': WeddingContactThumbnail, // kebab-case alias
+  gridGallery: ImageThumbnail, // reuse image thumbnail
+  'grid-gallery': ImageThumbnail, // kebab-case alias
+  slideGallery: ImageThumbnail, // reuse image thumbnail
+  'slide-gallery': ImageThumbnail, // kebab-case alias
+  music: DefaultThumbnail, // use default for music
+  default: DefaultThumbnail
+};
+
+export {
+  ButtonThumbnail,
+  TextThumbnail,
+  LinkThumbnail,
+  AttendThumbnail,
+  DdayThumbnail,
+  CalendarThumbnail,
+  BankAccountThumbnail,
+  CommentThumbnail,
+  ImageThumbnail,
+  MapThumbnail,
+  WeddingContactThumbnail,
+  DefaultThumbnail
+};
